feat: 세션, BaseController 수정

+++ prix/homepage/src/main/java/com/prix/homepage/backend/user/domain/User.java
@@ -10,9 +10,9 @@ import lombok.*;
 public class User {
 
 
-    private Long id;
+    private int id;
 
-    private Integer userid;
+    private int userid;
     private String email;
     private String password;
     private String name;
